refactor: move private test helpers to utils module to be used by many modules
