+++ parquet_file/src/lib.rs
@@ -12,3 +12,4 @@ pub mod chunk;
 pub mod metadata;
 pub mod storage;
 pub mod table;
+pub mod utils;
